+++ Change Log.md
@@ -40,6 +40,7 @@
 * Added ProgressUpdater
 * Added cancellation support in the async pipeline builder
 * Updated IElementDictionary to be enumerable, provide Names and Values, and the ability to change values within the json helper
+* Added DisposeWith() for any IAsyncDisposable, using CompositeAsyncDisposable
 
 
 ## AllOverIt.AspNetCore